refactor: add flamingock-cloud-api with basic requests/responses
