+++ cloud/flamingock-cloud-api/build.gradle.kts
@@ -1,0 +1,10 @@
+dependencies {
+}
+
+description = "Cloud Edition public API definitions"
+
+java {
+    toolchain {
+        languageVersion.set(JavaLanguageVersion.of(8))
+    }
+}
