+++ cloud/flamingock-cloud-api/src/main/java/io/flamingock/cloud/api/request/ClientSubmission.java
@@ -13,8 +13,18 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.flamingock.internal.common.cloud.vo;
+package io.flamingock.cloud.api.request;
 
-public enum TargetSystemAuditMarkType {
-    NONE, APPLIED, ROLLBACK
+import java.util.List;
+
+public class ClientSubmission {
+    private final List<StageRequest> stages;
+
+    public ClientSubmission(List<StageRequest> stages) {
+        this.stages = stages;
+    }
+
+    public List<StageRequest> getStages() {
+        return stages;
+    }
 }
