+++ cloud/flamingock-cloud-api/src/main/java/io/flamingock/cloud/api/request/StageRequest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.flamingock.internal.common.cloud.planner.request;
+package io.flamingock.cloud.api.request;
 
 import java.util.List;
