+++ cloud/flamingock-cloud-api/src/main/java/io/flamingock/cloud/api/request/TaskRequest.java
@@ -13,10 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.flamingock.internal.common.cloud.planner.request;
+package io.flamingock.cloud.api.request;
 
 
-import io.flamingock.internal.common.cloud.vo.TargetSystemAuditMarkType;
+import io.flamingock.cloud.api.vo.TargetSystemAuditMarkType;
 
 //TODO add recoveryStrategy, so we can determin the acction in the server
 public class TaskRequest {
