+++ cloud/flamingock-cloud-api/src/main/java/io/flamingock/cloud/api/response/StageResponse.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.flamingock.internal.common.cloud.planner.response;
+package io.flamingock.cloud.api.response;
 
 import java.util.List;
