+++ cloud/flamingock-cloud-api/src/main/java/io/flamingock/cloud/api/response/TaskResponse.java
@@ -13,18 +13,33 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.flamingock.internal.common.cloud.planner.request;
+package io.flamingock.cloud.api.response;
 
-import java.util.List;
+public class TaskResponse {
+    private String id;
+    private CloudChangeAction action;
 
-public class ClientSubmission {
-    private final List<StageRequest> stages;
-
-    public ClientSubmission(List<StageRequest> stages) {
-        this.stages = stages;
+    public TaskResponse() {
     }
 
-    public List<StageRequest> getStages() {
-        return stages;
+    public TaskResponse(String id, CloudChangeAction action) {
+        this.id = id;
+        this.action = action;
+    }
+
+    public String getId() {
+        return id;
+    }
+
+    public void setId(String id) {
+        this.id = id;
+    }
+
+    public CloudChangeAction getAction() {
+        return action;
+    }
+
+    public void setAction(CloudChangeAction action) {
+        this.action = action;
     }
 }
