+++ cloud/flamingock-cloud-api/src/main/java/io/flamingock/cloud/api/vo/TargetSystemAuditMarkType.java
@@ -13,8 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.flamingock.internal.common.cloud.vo;
+package io.flamingock.cloud.api.vo;
 
-public enum ActionResponse {
-    CONTINUE, EXECUTE, AWAIT
+public enum TargetSystemAuditMarkType {
+    NONE, APPLIED, ROLLBACK
 }
