+++ core/flamingock-core-commons/build.gradle.kts
@@ -2,6 +2,7 @@ val jacksonVersion = "2.16.0"
 val generalUtilVersion: String by extra
 val coreApiVersion: String by extra
 dependencies {
+    api(project(":cloud:flamingock-cloud-api"))
     api("io.flamingock:flamingock-core-api:${coreApiVersion}")
     api("io.flamingock:flamingock-general-util:${generalUtilVersion}")//todo implementation
     api("jakarta.annotation:jakarta.annotation-api:2.1.1")//todo can this be implementation?
